chore(lints): 清除 clippy deny 级欠账，恢复门禁
- panel.rs 两处 filter 冗余闭包改函数引用
- source.rs 删除恒等 f32::from 转换
- preset_menu.rs 消除中间 collect，filter 后直接 partition
- 5 处 GPUI 构建器签名补 #[expect(clippy::too_many_arguments, reason)]
- preset_state.rs 测试改结构体更新语法，清 field_reassign_with_default

cargo clippy -p frame-core -p frame-app --all-targets 0 error；
core 147 过/0 红，app 755 过/6 红既有（app_persistence）

+++ frame-app/src/app/settings_panel/source.rs
@@ -800,7 +800,7 @@ fn settings_bitrate_curve(
         .max()
         .unwrap_or(2) as f32)
         .mul_add(6.0, 6.0)
-        .clamp(16.0, f32::from(BITRATE_AXIS_LABEL_WIDTH_PX));
+        .clamp(16.0, BITRATE_AXIS_LABEL_WIDTH_PX);
     let mut y_axis = div()
         .relative()
         .flex_none()
